// src/config/config.type.ts
import { AppConfig } from './app-config.type';
import { AppleConfig } from '../auth-apple/config/apple-config.type';
import { AuthConfig } from '../auth/config/auth-config.type';
import { DatabaseConfig } from '../database/config/database-config.type';
import { FacebookConfig } from '../auth-facebook/config/facebook-config.type';
import { FileConfig } from '../files/config/file-config.type';
import { GoogleConfig } from '../auth-google/config/google-config.type';
import { MailConfig } from '../mail/config/mail-config.type';
import { TwitterConfig } from '../auth-twitter/config/twitter-config.type';
import { TwilioConfig } from '../twilio/config/twilio-config.type';
import { KafkaConfig } from 'src/kafka/config/kafka-config.type';
import { BridgerPayConfig } from 'src/transaction/services/bridgerpay/config/bridgerpay-config.type';
import { BinanceConfig } from 'src/transaction/services/binance/config/binance-config.type';
import { FreshDeskConfig } from 'src/fresh-desk/config/freshdesk-config.type';
import { ElasticsearchConfig } from './elastic-search.type';
import { RedisConfig } from 'src/redis/config/redis-config.type';
import { LegacyAPIConfig } from 'src/transaction/services/legacy/config/legacy-config.type';
import { NGeniusConfig } from 'src/transaction/services/n-genius/config/n-genius-config.type';
import { DpoConfig } from 'src/transaction/services/dpo/config/dpo-config.type';
import { PraxisConfigType } from 'src/transaction/services/praxis/config/praxis-config.type';
import { MyFatoorahConfig } from 'src/transaction/services/my-fatoorah/config/my-fatoorah-config.type';
import { AlphaPayConfig } from 'src/transaction/services/alphapay/config/alphapay-config.type';
import { PlugitConfigType } from 'src/plugit/config/plugit.config.type';
import { JenaPayConfig } from 'src/transaction/services/jenapay/config/jenapay-config.type';
import { IBAutomationConfig } from 'src/ib-automation/config/ib-automation-config.type';
import { ReferralConfig } from 'src/referral-program/config/referral-config.type';
import { ShuftiConfig } from 'src/kyc-shufti/config/shufti-config.type';
import { FirebaseConfig } from 'src/firebase/config/firebase-config.type';
import { PriceHistoryConfig } from 'src/mt5-manager/price-history/config/price-history.config-type';

export type AllConfigType = {
  app: AppConfig;
  apple: AppleConfig;
  auth: AuthConfig;
  database: DatabaseConfig;
  facebook: FacebookConfig;
  file: FileConfig;
  google: GoogleConfig;
  mail: MailConfig;
  twitter: TwitterConfig;
  twilio: TwilioConfig;
  kafka: KafkaConfig;
  bridgerpay: BridgerPayConfig;
  binance: BinanceConfig;
  freshdesk: FreshDeskConfig;
  elasticsearch: ElasticsearchConfig;
  redis: RedisConfig;
  legacy: LegacyAPIConfig;
  nGenius: NGeniusConfig;
  myFatoorah: MyFatoorahConfig;
  dpo: DpoConfig;
  praxis: PraxisConfigType;
  alphaPay: AlphaPayConfig;
  plugit: PlugitConfigType;
  jenapay: JenaPayConfig;
  ibAutomationConfig: IBAutomationConfig;
  referral: ReferralConfig;
  shuftiConfig: ShuftiConfig;
  firebase: FirebaseConfig;
  priceHistory: PriceHistoryConfig; // Migrated by Arshad Shaheen
};